+++ RELEASENOTES.md
@@ -1,1 +1,1 @@
-
+1. Update AWS CDK diff parser to account for different Stack name and construct id.